Moved PrepareAssemblies from package agc to cli
